Conversatonal Bot/requirements.txt
openai
streamlit
langchain_community
uvicorn